feat: separate t3rn ABI to separate module, fix deps
