+++ pallets/xdns/Cargo.toml
@@ -31,6 +31,7 @@ sp-version = { default-features = false, git = "https://github.com/paritytech/su
 pallet-timestamp = { default-features = false, git = "https://github.com/paritytech/substrate", branch = 'polkadot-v0.9.27' }
 t3rn-primitives  = { default-features = false, path = "../../primitives" }
 t3rn-types  = { default-features = false, path = "../../types", features = [ "runtime" ] }
+t3rn-abi  = { default-features = false, path = "../../types/abi", features = [ "runtime" ] }
 
 [dev-dependencies]
 circuit-mock-runtime    = { path = "../../runtime/mock" }
@@ -59,5 +60,6 @@ std = [
   "sp-std/std",
   "t3rn-primitives/std",
   "t3rn-types/std",
+  "t3rn-abi/std",
 ]
 try-runtime = [ "frame-support/try-runtime", "frame-system/try-runtime", "pallet-timestamp/try-runtime" ]
