+++ primitives/Cargo.toml
@@ -43,6 +43,7 @@ snowbridge-core = { path = "src/bridges/snowfork/core", default-features = false
 
 t3rn-sdk-primitives = { version = "=0.1.1-rc.4", default-features = false }
 t3rn-types          = { path = "../types", default-features = false, features = [ "runtime" ] }
+t3rn-abi            = { path = "../types/abi", default-features = false, features = [ "runtime" ] }
 
 [dev-dependencies]
 hex         = "0.4"
@@ -76,6 +77,7 @@ std = [
   "sp-finality-grandpa/std",
   "t3rn-sdk-primitives/std",
   "t3rn-types/std",
+  "t3rn-abi/std",
   "t3rn-types/runtime-std",
   "finality-grandpa/std",
   "pallet-grandpa-finality-verifier/std",
