+++ primitives/src/xdns.rs
@@ -8,10 +8,8 @@ use frame_system::pallet_prelude::OriginFor;
 use scale_info::TypeInfo;
 use sp_runtime::DispatchError;
 use sp_std::vec::Vec;
-use t3rn_types::{
-    fsx::SecurityLvl,
-    sfx::{SFXAbi, Sfx4bId},
-};
+use t3rn_abi::sfx_abi::SFXAbi;
+use t3rn_types::{fsx::SecurityLvl, sfx::Sfx4bId};
 
 pub type AllowedSideEffect = [u8; 4];
