+++ types/Cargo.toml
@@ -25,27 +25,19 @@ sp-runtime      = { git = "https://github.com/paritytech/substrate.git", branch 
 sp-std          = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true  }
 
 frame-system    = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true  }
-frame-support   = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27' }
-
-rlp = { version = "0.5.2", default-features = false, features = [ "derive" ] }
-ethabi = { version = "18.0.0", default-features = false, optional = true}
-ethereum-types = { version = "0.14.0", default-features = false , optional = true}
 
 [dev-dependencies]
 hex         = "0.4"
 hex-literal = "0.3.1"
 serde_json  = "1.0.41"
-rlp-derive = "0.1.0"
 sp-core     = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27' }
 sp-runtime  = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27' }
-pallet-balances  = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27' }
-frame-system  = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27' }
 
 t3rn-sdk-primitives = { version = "=0.1.1-rc.4" }
 circuit-runtime-types = { path = "../runtime/common-types" }
 
 [features]
 default     = [ "std" ]
-runtime     = [ "ethabi", "ethereum-types", "sp-runtime", "sp-std", "sp-core", "sp-io", "frame-system" ]
-runtime-std = [ "sp-runtime/std", "sp-core/std", "sp-io/std", "frame-system/std", "rlp/std", "ethabi/std", "ethereum-types/std"]
-std         = [ "sp-std/std", "sp-runtime/std", "frame-system/std", "frame-support/std", "sp-core/std", "codec/std", "scale-info/std", "serde/std", "bytes/std", "num-traits/std" ]
+runtime     = [ "sp-runtime", "sp-std", "sp-core", "sp-io", "frame-system" ]
+runtime-std = [ "sp-runtime/std", "sp-core/std", "sp-io/std", "frame-system/std" ]
+std         = [ "sp-std/std", "sp-runtime/std", "frame-system/std", "sp-core/std", "codec/std", "scale-info/std", "serde/std", "bytes/std", "num-traits/std" ]
