+++ types/abi/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+authors     = [ "t3rn ltd. <team@t3rn.io>" ]
+description = "A crate that with t3rn ABI cross-codec, SCALE, RLP, and more"
+edition     = "2021"
+homepage    = "https://t3rn.io"
+license     = "Apache-2.0"
+name        = "t3rn-abi"
+repository  = "https://github.com/t3rn/t3rn/"
+version     = "1.0.0-rc.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+bytes      = { version = "1.2.1", default-features = false, optional = true }
+num-traits = { version = "0.2", default-features = false, optional = true }
+codec      = { package = "parity-scale-codec", version = "3", default-features = false, features = [ "derive" ], optional = true }
+num        = { version = "0.4.0", default-features = false, optional = true }
+scale-info = { version = "2.1.1", default-features = false, features = [ "derive" ], optional = true }
+serde      = { default-features = false, version = "1.0", optional = true, features = [ "derive" ] }
+
+primitive-types = { version = "0.11.1", default-features = false, features = [ "scale-info" ], optional = true }
+
+sp-core         = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true }
+sp-runtime      = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true }
+sp-std          = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true }
+
+frame-system    = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true }
+frame-support   = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27', default-features = false, optional = true }
+
+rlp             = { version = "0.5.2", default-features = false, features = [ "derive" ], optional = true }
+
+[dev-dependencies]
+hex         = "0.4"
+hex-literal = "0.3.1"
+serde_json  = "1.0.41"
+ethabi = "18.0.0"
+rlp-derive = "0.1.0"
+pallet-balances  = { git = "https://github.com/paritytech/substrate.git", branch = 'polkadot-v0.9.27' }
+
+t3rn-sdk-primitives = { version = "=0.1.1-rc.4" }
+circuit-runtime-types = { path = "../../runtime/common-types" }
+
+[features]
+default     = [ "std" ]
+std         = [ "primitive-types/std", "sp-std/std", "sp-runtime/std", "frame-system/std", "frame-support/std", "sp-core/std", "codec/std", "scale-info/std", "serde/std", "bytes/std", "num-traits/std", "rlp/std" ]
+runtime     = [ "primitive-types", "sp-std", "sp-runtime", "frame-system", "frame-support", "sp-core", "codec", "scale-info", "serde", "bytes", "num-traits", "rlp" ]
+runtime-std = [ "primitive-types/std", "sp-std/std", "sp-runtime/std", "frame-system/std", "frame-support/std", "sp-core/std", "codec/std", "scale-info/std", "serde/std", "bytes/std", "num-traits/std", "rlp/std" ]
+try-runtime = [ "frame-support/try-runtime", "frame-system/try-runtime" ]
