+++ types/abi/src/types.rs
@@ -1,5 +1,4 @@
 use sp_std::vec::Vec;
-
 pub type Data = Vec<u8>;
 pub type Bytes = Data;
 pub type Name = Data;
